Implemented Colision
This is the best collision detection to date. Partly because everything
is nicely segmented, and partly because i kept it simple.

Notes: The size of the colliders is directly based on the max speed as
defined in player controller. This prevents the player from ever
clipping through even at max speed, as we will always be off by just 1
pixel. However, it may be worth adding a pixel or 2 for safety.