Added new sprite and everything that goes with it
Added a zelda sprite and added animation.

Then i gave the colliders some vertical offset, as the perspective of
the sprite means that the top of it should overlap the walls some. This
threw up numerous bugs which i think i have fixed, the main being that
the smaller colliders meant the velocity "safe limit" on the
subcolliders wasnt enough. Meaning if you spaz out near the walls, you
can cause the wrong collider to be hit, setting you to the other side of
the wall. And if this room is on the edge of the map, this will cause an
overflow exception in the levels array.